functions/requirements.txt
azure-functions==1.21.0
msal==1.31.0
requests==2.32.3
pyodbc==5.2.0
pyyaml==6.0.2
jsonschema==4.23.0
azure-monitor-opentelemetry==1.6.4
